Add checking if scenario can be fully played and sandbox too
When it comes to a scenario game, the 'canPlayEvents' will check if
there are enough events in the scenario file to play it fully, meaning
that all events will and can be played. Moreover, if there is no event
at all in the scenario or no event in the first season, game won't be
played.
Concerning sandbox, the only condition is that there is no season with
no event.