Add related events possibility in scenario mode
When the player makes a choice who involves related events, these
event(s) will take place randomly in future events.
They are placed in order to play the whole scenario.
The selected season where a related event will take place is the next
season of the basically scenario ending season (without related events).
The selected year is between the current year (+1 if its the last
season) and the last possible year of the scenario.